+++ contracts/MultisigWallet/contracts/MultiSigWalletFactory.sol
@@ -4,11 +4,9 @@ pragma solidity ^0.4.15;
 import "./Factory.sol";
 import "./MultiSigWallet.sol";
 
-
 /// @title Multisignature wallet factory - Allows creation of multisig wallet.
 /// @author Stefan George - <stefan.george@consensys.net>
 contract MultiSigWalletFactory is Factory {
-
     /*
      * Public functions
      */
@@ -16,10 +14,7 @@ contract MultiSigWalletFactory is Factory {
     /// @param _owners List of initial owners.
     /// @param _required Number of required confirmations.
     /// @return Returns wallet address.
-    function create(address[] _owners, uint _required)
-        public
-        returns (address wallet)
-    {
+    function create(address[] _owners, uint256 _required) public returns (address wallet) {
         wallet = new MultiSigWallet(_owners, _required);
         register(wallet);
     }
