+++ contracts/MultisigWallet/contracts/MultiSigWalletWithDailyLimitFactory.sol
@@ -4,11 +4,9 @@ pragma solidity ^0.4.15;
 import "./Factory.sol";
 import "./MultiSigWalletWithDailyLimit.sol";
 
-
 /// @title Multisignature wallet factory for daily limit version - Allows creation of multisig wallet.
 /// @author Stefan George - <stefan.george@consensys.net>
 contract MultiSigWalletWithDailyLimitFactory is Factory {
-
     /*
      * Public functions
      */
@@ -17,10 +15,11 @@ contract MultiSigWalletWithDailyLimitFactory is Factory {
     /// @param _required Number of required confirmations.
     /// @param _dailyLimit Amount in wei, which can be withdrawn without confirmations on a daily basis.
     /// @return Returns wallet address.
-    function create(address[] _owners, uint _required, uint _dailyLimit)
-        public
-        returns (address wallet)
-    {
+    function create(
+        address[] _owners,
+        uint256 _required,
+        uint256 _dailyLimit
+    ) public returns (address wallet) {
         wallet = new MultiSigWalletWithDailyLimit(_owners, _required, _dailyLimit);
         register(wallet);
     }
